fix(desktop): fix session telemetry truncation beyond 6.6m and smooth chart scale

+++ desktop/src/commonMain/kotlin/com/vinnovateit/latch/ui/components/ThroughputChart.kt
@@ -55,7 +55,7 @@ internal fun ThroughputChart(
             history = history,
             width = size.width,
             height = size.height,
-            maxRate = animatedPeak.toLong(),
+            maxRate = animatedPeak,
         )
 
         if (!isAmoled) {
